+++ dsa/src/components.rs
@@ -68,7 +68,7 @@ impl Components {
 
     /// Check whether the components are valid
     #[must_use]
-    pub fn is_valid(&self) -> bool {
+    pub(crate) fn is_valid(&self) -> bool {
         *self.p() >= two()
             && *self.q() >= two()
             && *self.g() >= BigUint::one()